Finalizando projeto

+++ backend/src/app/controllers/UserController.js
@@ -1,5 +1,6 @@
 import { Op } from 'sequelize';
 import User from '../models/User';
+import bcrypt from 'bcryptjs';
 
 class UserController {
   async store(req, res) {
@@ -21,7 +22,7 @@ class UserController {
     const fieldList = ["name", "email", "login"]
     if (fieldList.includes(field)){
       const users = await User.findAll({
-        where: {
+        attributes:["login", "name", "last_name", "email", "status", "password_expiration_date", "access_group"],where: {
             [field]: {
                 [Op.iLike]: value
             }
@@ -33,6 +34,11 @@ class UserController {
   }
   async delete(req, res){
     const userId = req.params.id
+    const user = await User.findByPk(req.userId)
+    console.log(JSON.stringify(user))
+    if(user.access_group !== 'superadmin' ){
+      return res.status(401).json({error:"User lacks privilege"})
+    }
     const deletedUser = await User.destroy({
       where: {
         id: userId
@@ -44,15 +50,32 @@ class UserController {
     return res.status(200).json({message:"User deleted successfully"})
   }
   async update(req, res){
+    // Pega o id pela url
     const userId = req.params.id
+
+    // Compara o id com a pk e retorna o usuário que os dois valores forem correspondentes
     const user = await User.findByPk(userId)
+
     if (!user){
+      // Retorna um erro se o o id e a pk não forem correspondentes
       return res.status(404).json({error:"User not exists"})
     }
+
+    // Cria uma variavel que receberá como valor a data/hora atual no momento da criação
+    let newDate = new Date()
+    
+    // Checa se o campo pass foi preenchido, sendo positivo, atualiza a expiration_date
+    if (req.body.password){
+      await user.update({password_expiration_date: newDate.setMonth(newDate.getMonth() + 3)})
+    }
+    // codigo 200 = sucesso
     let status = 200
+
+    // As informações além da expiration date serão atualizadas ou um erro será retornado
     try {
       await user.update(req.body)
     } catch (error) {
+      // codigo 400 = erro
       status = 400
     }
     return res.status(status).json({message: status == 200?"User updated successfully":"Database error"})
@@ -63,9 +86,21 @@ class UserController {
     if (!user){
       return res.status(404).json({error:"User not exists"})
     }
-    await user.update({password:"novasenha2024"})
+    const newDate = new Date()
+    newDate.setMonth(newDate.getMonth() + 3)
+
+    await user.update({password:"novasenha2024", password_expiration_date: newDate})
     
     return res.status(200).json({message:"You new password has been generated", newPassword:"novasenha2024"})
+  }
+  async index(req, res){
+    const users = await User.findAll({attributes:["id","login", "name", "last_name", "email", "status", "password_expiration_date", "access_group"]})
+    return res.status(200).json(users)
+  }
+  async getById(req, res){
+    const userId = req.params.id
+    const user = await User.findByPk(userId, {attributes:["id","login", "name", "last_name", "email", "status", "password_expiration_date", "access_group"]})
+    return res.status(200).json(user)
   }
 };
 export default new UserController();
